requirements.txt
fastapi
uvicorn[standard]
requests
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
openai
python-multipart
pymongo
python-dotenv
pydantic
joblib
xgboost==1.7.6
scikit-learn
pandas
numpy
